--- spring-cloud-contract-spec/src/main/groovy/org/springframework/cloud/contract/spec/util/ContentType.groovy
@@ -1,37 +1,0 @@
-/*
- *  Copyright 2013-2016 the original author or authors.
- *
- *  Licensed under the Apache License, Version 2.0 (the "License");
- *  you may not use this file except in compliance with the License.
- *  You may obtain a copy of the License at
- *
- *       http://www.apache.org/licenses/LICENSE-2.0
- *
- *  Unless required by applicable law or agreed to in writing, software
- *  distributed under the License is distributed on an "AS IS" BASIS,
- *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  See the License for the specific language governing permissions and
- *  limitations under the License.
- */
-
-package org.springframework.cloud.contract.spec.util
-
-/**
- * Represents content type
- *
- * @since 1.1.0
- */
-enum ContentType {
-
-	JSON("application/json"),
-	XML("application/xml"),
-	TEXT("text/plain"),
-	UNKNOWN("application/octet-stream")
-
-	final String mimeType
-
-	ContentType(String mimeType) {
-		this.mimeType = mimeType
-	}
-
-}
