+++ spring-cloud-contract-tools/spring-cloud-contract-spec-pact/pom.xml
@@ -19,7 +19,7 @@
 		</dependency>
 		<dependency>
 			<groupId>org.springframework.cloud</groupId>
-			<artifactId>spring-cloud-contract-spec</artifactId>
+			<artifactId>spring-cloud-contract-verifier</artifactId>
 		</dependency>
 		<dependency>
 			<groupId>org.springframework.boot</groupId>
@@ -51,11 +51,6 @@
 		<dependency>
 			<groupId>info.solidsoft.spock</groupId>
 			<artifactId>spock-global-unroll</artifactId>
-			<scope>test</scope>
-		</dependency>
-		<dependency>
-			<groupId>org.springframework.cloud</groupId>
-			<artifactId>spring-cloud-contract-verifier</artifactId>
 			<scope>test</scope>
 		</dependency>
 	</dependencies>
