+++ spring-cloud-contract-tools/spring-cloud-contract-spec-pact/src/main/groovy/org/springframework/cloud/contract/verifier/spec/converter/pact/PactContractConverter.groovy
@@ -21,7 +21,7 @@ import org.springframework.cloud.contract.spec.internal.ExecutionProperty
 import org.springframework.cloud.contract.spec.internal.Headers
 import org.springframework.cloud.contract.spec.internal.MatchingType
 import org.springframework.cloud.contract.spec.internal.QueryParameters
-import org.springframework.cloud.contract.spec.util.MapConverter
+import org.springframework.cloud.contract.verifier.util.MapConverter
 
 /**
  * Converter of JSON PACT file
