+++ spring-cloud-contract-verifier/src/main/groovy/org/springframework/cloud/contract/verifier/builder/RequestProcessingMethodBodyBuilder.groovy
@@ -28,12 +28,12 @@ import org.springframework.cloud.contract.spec.internal.NamedProperty
 import org.springframework.cloud.contract.spec.internal.QueryParameter
 import org.springframework.cloud.contract.spec.internal.Response
 import org.springframework.cloud.contract.spec.internal.Url
-import org.springframework.cloud.contract.spec.util.MapConverter
+import org.springframework.cloud.contract.verifier.util.MapConverter
 import org.springframework.cloud.contract.verifier.config.ContractVerifierConfigProperties
-import org.springframework.cloud.contract.spec.util.ContentType
+import org.springframework.cloud.contract.verifier.util.ContentType
 
-import static org.springframework.cloud.contract.spec.util.ContentUtils.recognizeContentTypeFromContent
-import static org.springframework.cloud.contract.spec.util.ContentUtils.recognizeContentTypeFromHeader
+import static org.springframework.cloud.contract.verifier.util.ContentUtils.recognizeContentTypeFromContent
+import static org.springframework.cloud.contract.verifier.util.ContentUtils.recognizeContentTypeFromHeader
 
 /**
  * An abstraction for creating a test method that includes processing of an HTTP request
