+++ spring-cloud-contract-verifier/src/main/groovy/org/springframework/cloud/contract/verifier/dsl/wiremock/BaseWireMockStubStrategy.groovy
@@ -21,9 +21,9 @@ import groovy.transform.PackageScope
 import groovy.transform.TypeChecked
 import org.springframework.cloud.contract.spec.internal.Headers
 import org.springframework.cloud.contract.spec.internal.DslProperty
-import org.springframework.cloud.contract.spec.util.MapConverter
-import org.springframework.cloud.contract.spec.util.ContentType
-import org.springframework.cloud.contract.spec.util.ContentUtils
+import org.springframework.cloud.contract.verifier.util.MapConverter
+import org.springframework.cloud.contract.verifier.util.ContentType
+import org.springframework.cloud.contract.verifier.util.ContentUtils
 
 import static ContentUtils.extractValue
 import static MapConverter.transformValues
@@ -51,7 +51,7 @@ abstract class BaseWireMockStubStrategy {
 	}
 
 	/**
-	 * For the given {@link org.springframework.cloud.contract.spec.util.ContentType} returns the String version of the body
+	 * For the given {@link ContentType} returns the String version of the body
 	 */
 	String parseBody(Object value, ContentType contentType) {
 		return parseBody(value.toString(), contentType)
