+++ spring-cloud-contract-verifier/src/main/groovy/org/springframework/cloud/contract/verifier/util/ContentUtils.groovy
@@ -15,19 +15,19 @@
  */
 
 package org.springframework.cloud.contract.verifier.util
+
 import groovy.json.JsonException
 import groovy.json.JsonOutput
 import groovy.json.JsonSlurper
 import groovy.transform.TypeChecked
 import groovy.util.logging.Slf4j
-import org.springframework.cloud.contract.spec.internal.Headers
 import org.codehaus.groovy.runtime.GStringImpl
-import org.springframework.cloud.contract.spec.internal.NamedProperty
 import org.springframework.cloud.contract.spec.internal.DslProperty
 import org.springframework.cloud.contract.spec.internal.ExecutionProperty
+import org.springframework.cloud.contract.spec.internal.Headers
 import org.springframework.cloud.contract.spec.internal.MatchingStrategy
+import org.springframework.cloud.contract.spec.internal.NamedProperty
 import org.springframework.cloud.contract.spec.internal.OptionalProperty
-import org.springframework.cloud.contract.spec.util.MapConverter
 
 import java.util.regex.Matcher
 import java.util.regex.Pattern
@@ -37,11 +37,12 @@ import static org.apache.commons.lang3.StringEscapeUtils.escapeJson
 import static org.apache.commons.lang3.StringEscapeUtils.escapeXml11
 
 /**
- * @deprecated use {@link org.springframework.cloud.contract.spec.util.ContentUtils}
+ * A utility class that can operate on a message body basing on the provided Content Type.
+ *
+ * @since 1.0.0
  */
 @TypeChecked
 @Slf4j
-@Deprecated
 class ContentUtils {
 
 	public static final Closure GET_STUB_SIDE = {
@@ -71,7 +72,7 @@ class ContentUtils {
 	 * @param valueProvider - provider of values either for server or client side
 	 * @return JSON structure with replaced client / server side parts
 	 */
-	public static Object extractValue(GString bodyAsValue, ContentType contentType, Closure valueProvider) {
+	static Object extractValue(GString bodyAsValue, ContentType contentType, Closure valueProvider) {
 		if (bodyAsValue.isEmpty()){
 			return bodyAsValue
 		}
@@ -100,7 +101,7 @@ class ContentUtils {
 		}
 	}
 
-	public static ContentType getClientContentType(GString bodyAsValue) {
+	static ContentType getClientContentType(GString bodyAsValue) {
 		try {
 			extractValueForJSON(bodyAsValue, GET_STUB_SIDE)
 			return ContentType.JSON
@@ -115,7 +116,7 @@ class ContentUtils {
 		}
 	}
 
-	public static ContentType getClientContentType(String bodyAsValue) {
+	static ContentType getClientContentType(String bodyAsValue) {
 		try {
 			new JsonSlurper().parseText(bodyAsValue)
 			return ContentType.JSON
@@ -129,11 +130,11 @@ class ContentUtils {
 		}
 	}
 
-	public static ContentType getClientContentType(Object bodyAsValue) {
+	static ContentType getClientContentType(Object bodyAsValue) {
 		return ContentType.UNKNOWN
 	}
 
-	public static ContentType getClientContentType(Map bodyAsValue) {
+	static ContentType getClientContentType(Map bodyAsValue) {
 		try {
 			JsonOutput.toJson(bodyAsValue)
 			return ContentType.JSON
@@ -142,7 +143,7 @@ class ContentUtils {
 		}
 	}
 
-	public static ContentType getClientContentType(List bodyAsValue) {
+	static ContentType getClientContentType(List bodyAsValue) {
 		try {
 			JsonOutput.toJson(bodyAsValue)
 			return ContentType.JSON
@@ -158,7 +159,7 @@ class ContentUtils {
 		)
 	}
 
-	public static Object extractValue(GString bodyAsValue, Closure valueProvider) {
+	static Object extractValue(GString bodyAsValue, Closure valueProvider) {
 		return extractValue(bodyAsValue, ContentType.UNKNOWN, valueProvider)
 	}
 
@@ -280,7 +281,7 @@ class ContentUtils {
 		return val[1]
 	}
 
-	public static ContentType recognizeContentTypeFromHeader(Headers headers) {
+	static ContentType recognizeContentTypeFromHeader(Headers headers) {
 		String content = headers?.entries.find { it.name == "Content-Type" } ?.clientValue?.toString()
 		if (content?.endsWith("json")) {
 			return ContentType.JSON
@@ -294,7 +295,7 @@ class ContentUtils {
 		return ContentType.UNKNOWN
 	}
 
-	public static MatchingStrategy.Type getEqualsTypeFromContentType(ContentType contentType) {
+	static MatchingStrategy.Type getEqualsTypeFromContentType(ContentType contentType) {
 		switch (contentType) {
 			case ContentType.JSON:
 				return MatchingStrategy.Type.EQUAL_TO_JSON
@@ -304,7 +305,7 @@ class ContentUtils {
 		return MatchingStrategy.Type.EQUAL_TO
 	}
 
-	public static ContentType recognizeContentTypeFromContent(GString gstring) {
+	static ContentType recognizeContentTypeFromContent(GString gstring) {
 		if (isJsonType(gstring)) {
 			return ContentType.JSON
 		}
@@ -314,15 +315,15 @@ class ContentUtils {
 		return ContentType.UNKNOWN
 	}
 
-	public static ContentType recognizeContentTypeFromContent(Map jsonMap) {
+	static ContentType recognizeContentTypeFromContent(Map jsonMap) {
 		return ContentType.JSON
 	}
 
-	public static ContentType recognizeContentTypeFromContent(List jsonList) {
+	static ContentType recognizeContentTypeFromContent(List jsonList) {
 		return ContentType.JSON
 	}
 
-	public static ContentType recognizeContentTypeFromContent(String string) {
+	static ContentType recognizeContentTypeFromContent(String string) {
 		try {
 			new JsonSlurper().parseText(string)
 			return ContentType.JSON
@@ -331,11 +332,11 @@ class ContentUtils {
 		}
 	}
 
-	public static ContentType recognizeContentTypeFromContent(Object gstring) {
+	static ContentType recognizeContentTypeFromContent(Object gstring) {
 		return ContentType.UNKNOWN
 	}
 
-	public static boolean isJsonType(GString gstring) {
+	static boolean isJsonType(GString gstring) {
 		if (gstring.isEmpty()) {
 			return false
 		}
@@ -354,7 +355,7 @@ class ContentUtils {
 		return false
 	}
 
-	public static boolean isXmlType(GString gstring) {
+	static boolean isXmlType(GString gstring) {
 		GString stringWithoutValues = new GStringImpl(
 				gstring.values.collect({
 					it instanceof String || it instanceof GString ? it.toString() : escapeXml11(it.toString())
@@ -370,7 +371,7 @@ class ContentUtils {
 		return false
 	}
 
-	public static ContentType recognizeContentTypeFromMatchingStrategy(MatchingStrategy.Type type) {
+	static ContentType recognizeContentTypeFromMatchingStrategy(MatchingStrategy.Type type) {
 		switch (type) {
 			case MatchingStrategy.Type.EQUAL_TO_XML:
 				return ContentType.XML
